include more tests, fix vcat/hcat
